[calibration] new calibration module
Move code from main_cameraCalibration to openMVG/calibration.
