+++ src/openMVG/CMakeLists.txt
@@ -21,6 +21,9 @@ IF(OpenMVG_BUILD_VOCTREE)
   ADD_SUBDIRECTORY(localization)
   ADD_SUBDIRECTORY(voctree)
   ADD_SUBDIRECTORY(dataio)
+  if(OpenMVG_USE_OPENCV)
+    ADD_SUBDIRECTORY(calibration)
+  endif(OpenMVG_USE_OPENCV)
 ENDIF(OpenMVG_BUILD_VOCTREE)
